Move editor settings to `editor`
