+++ crates/settings_ui/Cargo.toml
@@ -13,6 +13,7 @@ path = "src/settings_ui.rs"
 
 [dependencies]
 command_palette_hooks.workspace = true
+editor.workspace = true
 feature_flags.workspace = true
 gpui.workspace = true
 project.workspace = true
